Refactor file_processor to module directory
